+++ src/shared/ui/FormButtons/FormButtons.tsx
@@ -1,76 +1,73 @@
-import React, { FC } from 'react'
-import { TravelButton } from '@/shared/ui/Button/Button'
-import { Button, Flex, Popconfirm } from 'antd'
-import { FaRegTrashAlt } from 'react-icons/fa'
-import { QuestionCircleOutlined } from '@ant-design/icons'
-import styles from './style.module.scss'
-import cn from 'classnames'
+import { Button } from '@/components/ui/button';
+import { QuestionCircleOutlined } from '@ant-design/icons';
+import { Flex, Popconfirm } from 'antd';
+import cn from 'classnames';
+import { FC } from 'react';
+import { FaRegTrashAlt } from 'react-icons/fa';
+import styles from './style.module.scss';
 
 export interface FormButtonsProps {
-  onClose: () => void
-  onAccept: () => void
-  isLoading?: boolean
-  isEdit?: boolean
-  deleteText?: string
-  onDelete?: () => void
-  className?: string
+    onClose: () => void;
+    onAccept: () => void;
+    isLoading?: boolean;
+    isEdit?: boolean;
+    deleteText?: string;
+    onDelete?: () => void;
+    className?: string;
 }
 
 export const FormButtons: FC<FormButtonsProps> = ({
-  onAccept,
-  onClose,
-  isLoading = false,
-  isEdit = false,
-  deleteText,
-  onDelete,
-  className,
+    onAccept,
+    onClose,
+    isLoading = false,
+    isEdit = false,
+    deleteText,
+    onDelete,
+    className,
 }: FormButtonsProps) => {
-  return (
-    <div className={cn(className, styles.container)}>
-      {isEdit && onDelete && (
-        <div className={styles.deleteContainer}>
-          <Flex justify="end">
-            <Popconfirm
-              title={deleteText}
-              description={'Вы уверены?'}
-              icon={<QuestionCircleOutlined style={{ color: 'red' }} />}
-              onConfirm={onDelete}
-              okText={'Удалить'}
-              okButtonProps={{ style: { background: 'red' } }}
-              cancelText={'Отмена'}
-              onCancel={e => {
-                e?.preventDefault()
-                e?.stopPropagation()
-              }}
-            >
-              <Button
-                aria-label={'Удалить бронь'}
-                type="text"
-                icon={<FaRegTrashAlt />}
-                style={{ color: 'red' }}
-              >
-                {deleteText}
-              </Button>
-            </Popconfirm>
-          </Flex>
+    return (
+        <div className={cn(className, styles.container)}>
+            {isEdit && onDelete && (
+                <div className={styles.deleteContainer}>
+                    <Flex justify="end">
+                        <Popconfirm
+                            title={deleteText}
+                            description={'Вы уверены?'}
+                            icon={<QuestionCircleOutlined style={{ color: 'red' }} />}
+                            onConfirm={onDelete}
+                            okText={'Удалить'}
+                            okButtonProps={{ style: { background: 'red' } }}
+                            cancelText={'Отмена'}
+                            onCancel={(e) => {
+                                e?.preventDefault();
+                                e?.stopPropagation();
+                            }}
+                        >
+                            <Button
+                                aria-label={'Удалить бронь'}
+                                variant="outline"
+                                className="text-red-500"
+                            >
+                                {deleteText}
+                                <FaRegTrashAlt />
+                            </Button>
+                        </Popconfirm>
+                    </Flex>
+                </div>
+            )}
+            <div className="grid grid-cols-2  gap-2">
+                <Button
+                    variant="outline"
+                    className=" text-red-500"
+                    onClick={onClose}
+                    disabled={isLoading}
+                >
+                    Отмена
+                </Button>
+                <Button onClick={onAccept} disabled={isLoading}>
+                    {isEdit ? 'Сохранить' : 'Добавить'}
+                </Button>
+            </div>
         </div>
-      )}
-      <Flex justify="space-between">
-        <TravelButton
-          style={{ color: 'red', borderColor: 'red' }}
-          size="s"
-          view="secondary"
-          label="Отмена"
-          onClick={onClose}
-          disabled={isLoading}
-        />
-        <TravelButton
-          size="s"
-          label={isEdit ? 'Сохранить' : 'Добавить'}
-          onClick={onAccept}
-          loading={isLoading}
-        />
-      </Flex>
-    </div>
-  )
-}
+    );
+};